30/07

+++ SkyExams/Controllers/Study_ResourceController.cs
@@ -22,133 +22,56 @@ namespace SkyExams.Controllers
 
         public ActionResult resourceScreen(int? id)
         {
-            Sys_User user = db.Sys_User.ToList().Find(u => u.SysUser_ID == id);
-            return View(user);
+            ViewData["userID"] = "" + id;
+            Sys_User forRole = db.Sys_User.ToList().Find(u => u.SysUser_ID == id);
+            ViewData["userRole"] = "" + forRole.User_Role_ID;
+            List<Plane_Type> planeTypes = db.Plane_Type.ToList();
+            
+            return View(planeTypes);
         }// returns resource screen
+
+        public FileContentResult getImg(int id)
+        {
+            byte[] byteArray = db.Plane_Type.Find(id).Plane_Image;
+            return byteArray != null
+                ? new FileContentResult(byteArray, "image/jpeg")
+                : null;
+        }
 
         public ActionResult themeScreen(int? id, int? typeId)
         {
             ViewData["userID"] = "" + id;
             Sys_User user = db.Sys_User.Find(id);
+            ViewData["userRole"] = "" + user.User_Role_ID;
             int userRole = Convert.ToInt32(user.User_Role_ID);
             List<Study_Resource> resourceList = new List<Study_Resource>();
             if(userRole == 1)
             {
-                if (typeId == 1)
+                ViewData["themeID"] = "" + typeId;
+                ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == typeId).Type_Description;
+                Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
+                List<Student_Resource> stuResource = db.Student_Resource.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
+                foreach (Student_Resource temp in stuResource)
                 {
-                    ViewData["themeID"] = "" + 1;
-                    ViewData["planeType"] = "Cessna 172";
-                    Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
-                    List<Student_Resource> stuResource = db.Student_Resource.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
-                    foreach(Student_Resource temp in stuResource)
+                    List<Study_Resource> tempResources = db.Study_Resource.ToList();
+                    foreach (Study_Resource teResource in tempResources)
                     {
-                        List<Study_Resource> tempResources = db.Study_Resource.ToList();
-                        foreach(Study_Resource teResource in tempResources)
+                        if (teResource.Study_Resource_ID == temp.Study_Resource_ID && teResource.Rating_ID == typeId)
                         {
-                            if(teResource.Study_Resource_ID == temp.Study_Resource_ID && teResource.Rating_ID == 1)
-                            {
-                                resourceList.Add(teResource);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(resourceList);
-                }// cessna 172
-                if (typeId == 2)
-                {
-                    ViewData["themeID"] = "" + 2;
-                    ViewData["planeType"] = "Cessna 172 RG";
-                    Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
-                    List<Student_Resource> stuResource = db.Student_Resource.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
-                    foreach (Student_Resource temp in stuResource)
-                    {
-                        List<Study_Resource> tempResources = db.Study_Resource.ToList();
-                        foreach (Study_Resource teResource in tempResources)
-                        {
-                            if (teResource.Study_Resource_ID == temp.Study_Resource_ID && teResource.Rating_ID == 2)
-                            {
-                                resourceList.Add(teResource);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(resourceList);
-                }// cessna 172 RG
-                if (typeId == 3)
-                {
-                    ViewData["themeID"] = "" + 3;
-                    ViewData["planeType"] = "Cherokee 140";
-                    Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
-                    List<Student_Resource> stuResource = db.Student_Resource.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
-                    foreach (Student_Resource temp in stuResource)
-                    {
-                        List<Study_Resource> tempResources = db.Study_Resource.ToList();
-                        foreach (Study_Resource teResource in tempResources)
-                        {
-                            if (teResource.Study_Resource_ID == temp.Study_Resource_ID && teResource.Rating_ID == 3)
-                            {
-                                resourceList.Add(teResource);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(resourceList);
-                }// cherokee 140
-                if (typeId == 4)
-                {
-                    ViewData["themeID"] = "" + 4;
-                    ViewData["planeType"] = "Piper Twin Comanche";
-                    Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
-                    List<Student_Resource> stuResource = db.Student_Resource.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
-                    foreach (Student_Resource temp in stuResource)
-                    {
-                        List<Study_Resource> tempResources = db.Study_Resource.ToList();
-                        foreach (Study_Resource teResource in tempResources)
-                        {
-                            if (teResource.Study_Resource_ID == temp.Study_Resource_ID && teResource.Rating_ID == 4)
-                            {
-                                resourceList.Add(teResource);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(resourceList);
-                }// twin commanche
-                else
-                {
-                    return View(resourceList);
-                }
+                            resourceList.Add(teResource);
+                        }
+                    }// inner for each
+                }// for each
+                return View(resourceList);
+                
             }// if user is a student
             else
             {
-                if (typeId == 1)
-                {
-                    ViewData["themeID"] = "" + 1;
-                    ViewData["planeType"] = "Cessna 172";
-                    resourceList = db.Study_Resource.ToList().FindAll(p => p.Rating_ID == 1);
-                    return View(resourceList);
-                }// cessna 172
-                if (typeId == 2)
-                {
-                    ViewData["themeID"] = "" + 2;
-                    ViewData["planeType"] = "Cessna 172 RG";
-                    resourceList = db.Study_Resource.ToList().FindAll(p => p.Rating_ID == 2);
-                    return View(resourceList);
-                }// cessna 172 RG
-                if (typeId == 3)
-                {
-                    ViewData["themeID"] = "" + 3;
-                    ViewData["planeType"] = "Cherokee 140";
-                    resourceList = db.Study_Resource.ToList().FindAll(p => p.Rating_ID == 3);
-                    return View(resourceList);
-                }// cherokee 140
-                if (typeId == 4)
-                {
-                    ViewData["themeID"] = "" + 4;
-                    ViewData["planeType"] = "Piper Twin Comanche";
-                    resourceList = db.Study_Resource.ToList().FindAll(p => p.Rating_ID == 4);
-                    return View(resourceList);
-                }// twin commanche
-                else
-                {
-                    return View(resourceList);
-                }
+                ViewData["themeID"] = "" + typeId;
+                ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == typeId).Type_Description;
+                resourceList = db.Study_Resource.ToList().FindAll(p => p.Rating_ID == typeId);
+                return View(resourceList);
+                
             }// if user is not a student
         }// theme screen
 
@@ -200,6 +123,7 @@ namespace SkyExams.Controllers
         {
             ViewData["loggedId"] = "" + loggedId;
             Study_Resource delResource = db.Study_Resource.ToList().Find(p => p.Study_Resource_ID == id);
+            ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == delResource.Rating_ID).Type_Description;
             return View(delResource);
         }// delete Resource
 
@@ -209,8 +133,11 @@ namespace SkyExams.Controllers
             db.Study_Resource.Remove(delResource);
             db.SaveChanges();
             Student_Resource delStuResource = db.Student_Resource.ToList().Find(r => r.Study_Resource_ID == delResource.Study_Resource_ID);
-            db.Student_Resource.Remove(delStuResource);
-            db.SaveChanges();
+            if(delStuResource != null)
+            {
+                db.Student_Resource.Remove(delStuResource);
+                db.SaveChanges();
+            }
             return RedirectToAction("resourceScreen", new { id = loggedId });
         }// delete conformation
